feat: Add deletePost mutation to allow users to delete their posts
This code change adds a new mutation, `deletePost`, to the resolvers.js file. The mutation allows authenticated users to delete their own posts by providing the post ID. An authentication check is performed to ensure that only the creator of the post can delete it. If the user is not authenticated or not the creator of the post, an appropriate error is thrown. This enhancement improves the functionality and security of the application by allowing users to delete their posts.