refactor: Update package.json and yarn.lock to use TypeScript 5.5.4
This code change updates the package.json and yarn.lock files to include the TypeScript dependency at version 5.5.4. This update ensures that the project uses the latest stable version of TypeScript, which may include bug fixes and performance improvements. By keeping the dependencies up to date, we can take advantage of the latest features and improvements provided by the TypeScript language.