feat: Add todos model and interface
This code change adds a new file, todos.ts, which contains the ITodo interface and the todos array. The ITodo interface defines the structure of a todo object, including the id and text properties. The todos array is an empty array that will store todo objects. This addition allows for the management of todos within the application.