```text refactor: Rename todos.js to todos.controller.js and update todos route
This code change renames the file todos.js to todos.controller.js in the models directory. It also updates the todos route in the routes directory to use the new todos.controller.js file. This refactor improves the code's organization and follows a more consistent naming convention for controller files.