refactor: Update server.js to enable HTTPS support
This code change updates the `server.js` file to enable HTTPS support for the application. It adds the necessary code to create an HTTPS server using the provided private key and certificate files. This enhancement improves the security of the application by enabling secure communication with clients over an encrypted connection.