refactor: Update server.js to listen on '0.0.0.0' for improved network accessibility
This code change updates the `server.js` file to modify the `app.listen` function to listen on the IP address '0.0.0.0' in addition to the specified port. This improvement allows the server to accept connections from any network interface, making the application accessible from external devices. This modification enhances the network accessibility of the application, enabling it to be accessed by clients on different networks or devices.